Remove todoist from projects
Before everything was handled as if a project and then become a mobile
task, but now for non-enumerated sections (aka exceptions), starts as a
mobile task that becomes a project.

Allows for more freq checking of justification from a skippable daily to
every single task as this is where the exception tasks are in general
completed in.